+++ Assets/Scripts/IdleMiner/Models/DataModels/TransporterRemoteData.cs
@@ -18,7 +18,7 @@ namespace PG.IdleMiner.Models.DataModels
         public double LoadedCash;
 
         [JsonProperty("CurrentLocation")]
-        public int CurrentLocation;
+        public float CurrentLocation;
 
         [JsonProperty("TransporterState")]
         public ETransporterState TransporterState;
